requirements.txt
Django==3.2.7
django-crispy-forms==1.12.0
django-ajax-helpers==0.0.22
django-tab-menus==0.0.16
django-nested-modals==0.0.23
celery
redis
